requirements.txt
pandas>=2.0.0
requests>=2.31.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
PyYAML>=6.0.1
pytest>=7.4.0
